refactor magcal & friends into libcalib (linux&win)
- many declarations moved to libcalib.h from imuread.h
- a few rawdata (libcalib) routines moved to serialdata (app)
- apply_calibration moved to libcalib.cpp from app's visualise.cpp
- msvc and mingw working
- linux working
- mac not tested
