+++ libcalib/mahony.c
@@ -20,7 +20,7 @@
 
 //----------------------------------------------------------------------------------------------
 
-#include "imuread.h"
+#include "libcalib.h"
 
 #ifdef USE_MAHONY_FUSION
