+++ libcalib/quality.c
@@ -1,4 +1,7 @@
-#include "imuread.h"
+#include "libcalib.h"
+
+#include <string.h>
+#include <math.h>
 
 // Discussion of what these 4 quality metrics really do
 // https://forum.pjrc.com/threads/59277-Motion-Sensor-Calibration-Tool-Parameter-Understanding
