+++ msvc/2017/MotionCal.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 15.0.28307.1267
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "MotionCal", "MotionCal.vcxproj", "{CAA79469-08CA-422F-B1F0-568191FCBA3A}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "libcalib", "libcalib\libcalib.vcxproj", "{0A49EE6A-1DB1-48C8-93E2-5355B47783E9}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
@@ -21,6 +23,14 @@ Global
 		{CAA79469-08CA-422F-B1F0-568191FCBA3A}.Release|x64.Build.0 = Release|x64
 		{CAA79469-08CA-422F-B1F0-568191FCBA3A}.Release|x86.ActiveCfg = Release|Win32
 		{CAA79469-08CA-422F-B1F0-568191FCBA3A}.Release|x86.Build.0 = Release|Win32
+		{0A49EE6A-1DB1-48C8-93E2-5355B47783E9}.Debug|x64.ActiveCfg = Debug|x64
+		{0A49EE6A-1DB1-48C8-93E2-5355B47783E9}.Debug|x64.Build.0 = Debug|x64
+		{0A49EE6A-1DB1-48C8-93E2-5355B47783E9}.Debug|x86.ActiveCfg = Debug|Win32
+		{0A49EE6A-1DB1-48C8-93E2-5355B47783E9}.Debug|x86.Build.0 = Debug|Win32
+		{0A49EE6A-1DB1-48C8-93E2-5355B47783E9}.Release|x64.ActiveCfg = Release|x64
+		{0A49EE6A-1DB1-48C8-93E2-5355B47783E9}.Release|x64.Build.0 = Release|x64
+		{0A49EE6A-1DB1-48C8-93E2-5355B47783E9}.Release|x86.ActiveCfg = Release|Win32
+		{0A49EE6A-1DB1-48C8-93E2-5355B47783E9}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
